Preserve backup locators and close API contract regressions
